# docs/openapi/auth-service.openapi.yaml
openapi: 3.0.3
info:
  title: NebulaOps auth-service API
  version: 23.4.0
  description: Live-only backend API. Responses are produced by services/adapters invoking real APIs, sockets, repositories, or CLI tools. No mock/statistical sample data is returned.
servers:
  - url: http://localhost:8081
paths:
  /api/auth/register:
    post:
      summary: Create user account
      tags:
        - auth-service
      responses:
        '200':
          description: Live response from backend service.
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid request.
        '500':
          description: Runtime tool/API error.
  /api/auth/login:
    post:
      summary: Login user
      tags:
        - auth-service
      responses:
        '200':
          description: Live response from backend service.
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid request.
        '500':
          description: Runtime tool/API error.
...